Refactoring classes into individual Files
