+++ UnityEngineAnalyzer.CLI/Reporting/IAnalyzerExporter.cs
@@ -1,0 +1,11 @@
+﻿using System;
+
+namespace UnityEngineAnalyzer.CLI.Reporting
+{
+    public interface IAnalyzerExporter
+    {
+        void AppendDiagnostic(DiagnosticInfo diagnosticInfo);
+        void Finish(TimeSpan duration);
+        void InitializeExporter(string fileName);
+    }
+}
