+++ UnityEngineAnalyzer.CLI/UnityEngineAnalyzer.CLI.csproj
@@ -100,6 +100,10 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="AnalyzerReport.cs" />
+    <Compile Include="Reporting\JsonAnalyzerExporter.cs" />
+    <Compile Include="Reporting\ConsoleAnalyzerExporter.cs" />
+    <Compile Include="Reporting\DiagnosticInfo.cs" />
+    <Compile Include="Reporting\IAnalyzerExporter.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="SolutionAnalyzer.cs" />
